+++ demos/src/bin/shooter_like.rs
@@ -75,13 +75,13 @@ fn main() {
     {
         match app_setup_configuration.schedule_to_use {
             ScheduleToUse::Update => {
-                app.add_plugins(PhysicsPlugins::default());
+                app.add_plugins(PhysicsPlugins::new(PostUpdate));
                 // To use Tnua with avian3d, you need the `TnuaAvian3dPlugin` plugin from
                 // bevy-tnua-avian3d.
-                app.add_plugins(TnuaAvian3dPlugin::default());
+                app.add_plugins(TnuaAvian3dPlugin::new(Update));
             }
             ScheduleToUse::FixedUpdate => {
-                app.add_plugins(PhysicsPlugins::new(FixedUpdate));
+                app.add_plugins(PhysicsPlugins::new(FixedPostUpdate));
                 app.add_plugins(TnuaAvian3dPlugin::new(FixedUpdate));
             }
             ScheduleToUse::PhysicsSchedule => {
@@ -188,15 +188,11 @@ fn setup_player(mut commands: Commands, asset_server: Res<AssetServer>) {
     {
         cmd.insert(rapier::RigidBody::Dynamic);
         cmd.insert(rapier::Collider::capsule_y(0.5, 0.5));
-        // For Rapier, an "IO" bundle needs to be added so that Tnua will have all the components
-        // it needs to interact with Rapier.
-        cmd.insert(TnuaRapier3dIOBundle::default());
     }
     #[cfg(feature = "avian3d")]
     {
         cmd.insert(avian::RigidBody::Dynamic);
         cmd.insert(avian::Collider::capsule(0.5, 1.0));
-        // Avian does not need an "IO" bundle.
     }
 
     // `TnuaController` is Tnua's main interface with the user code. Read