+++ examples/example_animating.rs
@@ -16,15 +16,15 @@ fn main() {
         .add_plugins((
             DefaultPlugins,
             PhysicsPlugins::default(),
-            TnuaControllerPlugin::default(),
-            TnuaAvian3dPlugin::default(),
+            TnuaControllerPlugin::new(FixedUpdate),
+            TnuaAvian3dPlugin::new(FixedUpdate),
         ))
         .add_systems(
             Startup,
             (setup_camera_and_lights, setup_level, setup_player),
         )
         .add_systems(
-            Update,
+            FixedUpdate,
             (
                 apply_controls.in_set(TnuaUserControlsSystemSet),
                 prepare_animations,
@@ -119,7 +119,6 @@ fn setup_player(mut commands: Commands, asset_server: Res<AssetServer>) {
         // By locking the rotation we can prevent this.
         LockedAxes::ROTATION_LOCKED.unlock_rotation_y(),
     ));
-    // NOTE: if this was Rapier, we'd also need `TnuaRapier3dIOBundle`. Avian does not need it.
 }
 
 // No Tnua-related setup here - this is just for dealing with Bevy's animation graph.