+++ rapier3d/CHANGELOG.md
@@ -10,6 +10,11 @@ NOTE: This changelog is shared between bevy-tnua-rapier2d and bevy-tnua-rapier3d
 ### Deprecated
 - Deprecate `TnuaRapier3dIOBundle` in favor or bevy required components.
 
+### Fixed
+- Improve the cast-inside-self check (Fixes
+  https://github.com/idanarye/bevy-tnua/issues/85)
+  Note that it does cause https://github.com/idanarye/bevy-tnua/issues/87
+
 ## 0.9.0 - 2024-12-13
 ### Changed
 - Use `RapierContextEntityLink` to detect the Rapir context. This means Tnua