+++ rapier3d/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 NOTE: This changelog is shared between bevy-tnua-rapier2d and bevy-tnua-rapier3d.
 
 ## [Unreleased]
+### Changed
+- Upgrade to bevy_rapier 0.29.
+
 ### Deprecated
 - Deprecate `TnuaRapier3dIOBundle` in favor or bevy required components.